% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_slopes_and_mean.R
\name{filter_slopes_and_mean}
\alias{filter_slopes_and_mean}
\title{Filter slope and mean data to remove scarce data ranges before fitting polynomial}
\usage{
filter_slopes_and_mean(
  data,
  performance_metric = "mean",
  window_size,
  min_obs,
  max_filter = TRUE,
  min_filter = TRUE
)
}
\arguments{
\item{data}{output from \code{longpoly::get_slopes_and_mean()} (or any tibble with columns "id", "performance_slope", & "performance_mean" or "performance_bl")}

\item{performance_metric}{which measure of performance is being used? Must be either "mean" (in which case performance_mean column must be in data) or "baseline" (data must contain performance_bl). Default = \code{"mean"}}

\item{window_size}{the difference added to /subtracted from the minuimum/maximum values to subset the data. The number of observations in this subset is counted}

\item{min_obs}{the minimum number of observations required in the subset defined by \code{window_size}. If the fewer observations exist, data in these ranges are filtered out}

\item{max_filter}{filter highest scores? default = \code{TRUE}}

\item{min_filter}{filter lowest scores? default = \code{TRUE}}
}
\value{
a dataset containing slope and mean values with sparse data removed (if requested and identified)
}
\description{
Filter slope and mean data to remove scarce data ranges before fitting polynomial
}
\examples{

longpoly_example_data \%>\% nrow

filtered_example_data <- filter_slopes_and_mean(
data = longpoly_example_data,
performance_metric = "mean",
window_size = 0.25,
min_obs = 20,
max_filter = TRUE,
min_filter = TRUE
)

filtered_example_data \%>\% nrow
}
