% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_polynomial.R
\name{find_polynomial}
\alias{find_polynomial}
\title{A data-driven approach to choosing order of polynomial model describing the relationship between linear rate of change and mean score in longitudinal data}
\usage{
find_polynomial(
  data,
  performance_metric = "mean",
  test_proportion = 1/3,
  idcol = "id",
  max_order = 6,
  x
)
}
\arguments{
\item{data}{output from \code{longpoly::get_slopes_and_mean()} (or any tibble with columns "id", "performance_slope", & "performance_mean" or "performance_bl")}

\item{performance_metric}{which measure of performance is being used? Must be either "mean" (in which case performance_mean column must be in data) or "baseline" (data must contain performance_bl). Default = \code{"mean"}}

\item{test_proportion}{the proportion of participants to allocate to the test set. default = \code{1/3}}

\item{idcol}{the column name corresponding to the ID variable in data. Default = \code{"id"}}

\item{max_order}{the maximum order for polynomial models to be tested. Polynomial models for 1:\code{max_order} will be tested. default = \code{6}}

\item{x}{Polynomials are selected after finding the maximum proportion of variance explained (PVE) in test data across all models, and then finding the most parsimonious model with a PVE that is at least max(PVE) - x}
}
\value{
a list containing:
\enumerate{
\item 'polynomial_results' - a tibble with columns for the order of the model tested and the corresponding pve in the train and test data
\item 'selected_order' - the order of the selected model (based on the max(PVE) - x criteria)
\item 'selected_model_coefficients' - coefficients of the selected model (based on the max(PVE) - x criteria)
\item 'train_ids' — a character vector of ids allocated to the train dataset in model development
\item 'test_ids' — a character vector of ids allocated to the test dataset in model development
}
}
\description{
A data-driven approach to choosing order of polynomial model describing the relationship between linear rate of change and mean score in longitudinal data
}
\examples{
find_poly_results <- find_polynomial(data = longpoly_example_data,
                                     performance_metric = "mean",
                                     idcol = "id",
                                     test_proportion = 1/3,
                                     max_order = 6,
                                     x = 0.05)

# View proportion of variance explained in the train and test data for each polynomial tested
find_poly_results$polynomial_results

# View the order of the selected model
find_poly_results$selected_order

# View the coefficients of the selected model
find_poly_results$selected_model_coefficients

# The IDs of the participants allocated to the train and test sets are stored in the following vectors
find_poly_results$train_ids |> head()
find_poly_results$test_ids |> head()
}
