% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_polynomial.R
\name{implement_polynomial}
\alias{implement_polynomial}
\title{Fit a polynomial of a specified order to describe the relationship between linear rate of change and mean score in longitudinal data}
\usage{
implement_polynomial(
  data,
  performance_metric = "mean",
  order = 3,
  floor_effects = FALSE,
  floor_range = NULL
)
}
\arguments{
\item{data}{output from \code{longpoly::get_slopes_and_mean()} (or any tibble with columns "id", "performance_slope", & "performance_mean" or "performance_bl")}

\item{performance_metric}{which measure of performance is being used? Must be either "mean" (in which case performance_mean column must be in data) or "baseline" (data must contain performance_bl). Default = \code{"mean"}}

\item{order}{the order of the polynomial model (recommended to be selected on the scree plot output from \code{longpoly::test_polynomial()}). default = \code{3}}

\item{floor_effects}{test for floor effects? These are identified by finding the mean value of the performance variable that is associated with the fastest rate of decline in the impaired range (specified by floor_range()). default = \code{FALSE}}

\item{floor_range}{(required if \code{floor_effects = TRUE}). the 'impaired' range over which to search for the value of "performance_mean" that returns the minimum value for "performance_slope". expects a vector of length two with the first value providing the lower bound of the range and the second value the upper.}
}
\value{
a list containing:
\enumerate{
\item \code{model_formula} — the formula of the final model
\item \code{final_data} — a tibble with columns for "id", "performance_slope", either "performance_mean" or "performance_bl" (depending on \code{performance_metric}), "predicted_slope", and "residual." If \code{floor_effects = TRUE}, an additional column "floor_effects" is appended with either keep/remove reflecting whether the record should be removed based on the identified performance_mean threshold for floor effects
\item \code{threshold} — if \code{floor_effects = TRUE}, this records the threshold used for floor effect classifications
}
}
\description{
Fit a polynomial of a specified order to describe the relationship between linear rate of change and mean score in longitudinal data
}
\examples{

poly_out <-
implement_polynomial(
  data = longpoly_example_data,
  performance_metric = "mean",
  order = 3,
  floor_effects = TRUE,
  floor_range = c(min(longpoly_example_data$performance_mean), 0)
)

# View the equation of the final model
poly_out$model_formula

# View the final data
poly_out$final_data |> head()

# View the selected threshold
 poly_out$threshold

# Other example uses:

# implement_polynomial(example_slope_mean_data,
#                      floor_effects = FALSE)
#
# # if floor_effects = TRUE,
# it may be sensible to use the minumum performance_mean value as the lower bound of the range,
# e.g.:
# example_slope_mean_data <-
#   longpoly::get_slopes_and_mean(
#     example_data,
#     id_col = "id",
#     outcome_col = "memory_test",
#     time_col = "time"
#   ))
#
#
# implement_polynomial(
#   example_slope_mean_data,
#   floor_effects = TRUE,
#   floor_range = c(min(
#     example_slope_mean_data$performance_mean
#   ), 0)
# )
#
#
# # if floor_effects = FALSE,
# it is straightforward to pipe in the output from longpoly::get_slopes_and_mean(),
# e.g.:
# get_slopes_and_mean(example_data,
#                    id_col = "id",
#                    outcome_col = "memory_test",
#                    time_col = "time") \%>\%
#   implement_polynomial(., floor_effects = FALSE)

}
