% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longpoly_example_data}
\alias{longpoly_example_data}
\title{Longpoly Example Data}
\format{
\subsection{\code{longpoly_example_data}}{

A data frame with 2,000 rows and 3 columns:
\describe{
\item{id}{Participant ID}
\item{performance_mean}{Mean Value across Longitudinal Data (Simulated)}
\item{performance_slope}{Linear Slope across Longitudinal Data (Simulated)}
...
}
}
}
\source{
\url{https://github.com/shffer/longpoly?tab=readme-ov-file#a-note-about-simulated-data}
}
\usage{
longpoly_example_data
}
\description{
Simulated Data Used in Longpoly Examples
}
\keyword{datasets}
