% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_polynomial.R
\name{plot_polynomial}
\alias{plot_polynomial}
\title{Plot selected models from \code{final_data} tibble output from the \code{longpoly::implement_polynomial()} function}
\usage{
plot_polynomial(
  data,
  order = 3,
  x_label = "Mean Performance",
  y_label = "Slope",
  line_width = 1.5,
  whole_cohort_point_color = "#325a9c",
  whole_cohort_line_color = "#110036",
  whole_cohort_title = NULL,
  whole_cohort_only = TRUE,
  train_id = NULL,
  test_id = NULL,
  train_point_color = "#430C33",
  train_line_color = "#7B115B",
  train_title = NULL,
  test_point_color = "#113B19",
  test_line_color = "#7B115B",
  test_title = NULL,
  show_equation = TRUE,
  keep_remove = FALSE,
  threshold = NULL,
  threshold_linetype = "dashed",
  threshold_line_color = "#5f6a7a",
  x_offset = 0.25,
  y_offset = 0.35,
  remove_point_color = "grey",
  annotate_floor_thresh = FALSE,
  legend_position = "none",
  legend_title = "Floor Effects",
  floor_keep_label = "Keep",
  floor_remove_label = "Remove"
)
}
\arguments{
\item{data}{the \code{final_data} tibble output from the \code{longpoly::implement_polynomial()}}

\item{order}{the order of the polynomial model (recommended to be selected on the scree plot output from \code{longpoly::test_polynomial()}). default = \code{3}}

\item{x_label}{a character vector to optionally change the x axis label. default = "Mean Performance"}

\item{y_label}{a character vector to optionally change the y axis label. default = "Slope"}

\item{line_width}{the width of the line reflecting the polynomial equation. default = \code{1.5}}

\item{whole_cohort_point_color}{optionally set the color of the points in the whole cohort plot}

\item{whole_cohort_line_color}{optionally set the color of the line in the whole cohort plot}

\item{whole_cohort_title}{optionally set the title of the whole cohort plot. default = \code{NULL}}

\item{whole_cohort_only}{should only the whole cohort plot be generated? if \code{FALSE}, plots showing the train/test performance in \code{long_poly::test_polynomial()} for the selected order are returned. default = \code{TRUE}}

\item{train_id}{a vector of containing the IDs of observations allocated to the training dataset (recommended to use \code{train_ids} output from \code{long_poly::test_polynomial()})}

\item{test_id}{a vector of containing the IDs of observations allocated to the test dataset (recommended to use \code{test_ids} output from \code{long_poly::test_polynomial()})}

\item{train_point_color}{optionally set the color of the points in the training data plot}

\item{train_line_color}{optionally set the color of the line in the training data plot}

\item{train_title}{optionally set the title of the the training data plot. default = \code{NULL}}

\item{test_point_color}{optionally set the color of the points in the test data plot}

\item{test_line_color}{optionally set the color of the line in the test data plot}

\item{test_title}{optionally set the title of the test data plot. default = \code{NULL}}

\item{show_equation}{show the equation of the model in the plots? default = \code{TRUE}}

\item{keep_remove}{produce an additional plot showing IDs removed due to floor effects? default = \code{FALSE}}

\item{threshold}{(required if \code{keep_remove = TRUE}) the threshold used to keep or remove records due to floor_effects (recommended to use \code{threshold} output from \code{long_poly::implement_polynomial()})}

\item{threshold_linetype}{ggplot linetype arguments to control type of line when \code{keep_remove = TRUE}. set to "blank" to remove line}

\item{threshold_line_color}{optionally set color of threhold line when \code{keep_remove = TRUE}}

\item{x_offset}{increase or decrease the x axis plotting region. this value is subtracted and added from the min and max values of performance_mean, respectively, to specify the plotting region in relation to the observed data. default = 0.25}

\item{y_offset}{increase or decrease the y axis plotting region. this value is subtracted and added from the min and max values of performance_slope, respectively, to specify the plotting region in relation to the observed data. default = 0.35}

\item{remove_point_color}{optionally set the point color for removed participants due to floor effects when \code{keep_remove = TRUE} (those kept will be plotted with color specified in \code{whole_cohort_point_color})}

\item{annotate_floor_thresh}{include an annotation specifying the threshold for floor effects when \code{keep_remove = TRUE}}

\item{legend_position}{ggplot legend position argument to control the placement of a legend identifying the participants kept and removed when \code{keep_remove = TRUE}. default = "none"}

\item{legend_title}{if a legend position is set and \code{keep_remove = TRUE}, this controls the title of the legend. deault = "Floor Effects"}

\item{floor_keep_label}{if a legend position is set and \code{keep_remove = TRUE}, this controls the label of records with scores above the identified floor effects threshold. default = "Keep"}

\item{floor_remove_label}{if a legend position is set and \code{keep_remove = TRUE}, this controls the label of records with scores below the identified floor effects threshold. default = "Remove"}
}
\value{
a list of ggplot objects:
\enumerate{
\item \code{whole_cohort} — the selected polynomial applied to all records
\item \code{train} — (if requested) the selected polynomial applied to train data
\item \code{test} — (if requested) the selected polynomial applied to test data. note that the polynomial developed in the train data is applied to the test data
\item \code{keep_remove}  — (if requested) the whole_cohort plot with additional visualisation for the removal of records due to floor effects
}
}
\description{
Plot selected models from \code{final_data} tibble output from the \code{longpoly::implement_polynomial()} function
}
\examples{

# if implement_polynomial results were output to polynomial_data, plots can be produced as follows

#' # plots <-
#   plot_polynomials(
#     data = polynomial_data$final_data,
#     order = 3,
#     whole_cohort_title = "Selected Polynomial in Whole Cohort",
#     whole_cohort_only = FALSE,
#     train_id = results$train_ids,
#     test_id = results$test_ids,
#     train_title = "Selected Polynomial in Train Cohort",
#     test_title = "Selected Polynomial in Test Cohort",
#     keep_remove = TRUE,
#     threshold = polynomial_data$threshold,
#     threshold_linetype = "dashed",
#     threshold_line_color = "#5f6a7a",
#     annotate_floor_thresh = TRUE,
#     legend_position = "bottom"
#   )

# plots

}
