% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slopes_and_mean.R
\name{get_slopes_and_mean}
\alias{get_slopes_and_mean}
\title{Get individual linear slope and mean values from longitudinal data}
\usage{
get_slopes_and_mean(
  data,
  id_col = "id",
  outcome_col = "outcome",
  time_col = "time"
)
}
\arguments{
\item{data}{a longitudinal dataset}

\item{id_col}{the column name corresponding to the ID variable in data. Default = \code{"id"}}

\item{outcome_col}{the column name corresponding to the outcome variable in data. Default = \code{"outcome"}}

\item{time_col}{the column name corresponding to the time variable in data. Default = \code{"time"}}
}
\value{
a \code{tibble} with \code{id}, \code{performance_slope}, and \code{performance_mean} columns. This is what is used as inputs for subsequent \code{test_polynomial()} and \code{implement_polynomial()} function calls (and can therefore be directly piped in without storing intermediate output)
}
\description{
Get individual linear slope and mean values from longitudinal data
}
\examples{

# get_slopes_and_mean(example_data, id_col = "id", outcome_col = "memory_test", time_col = "time")

}
