% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slopes_and_mean.R
\name{get_slopes_and_mean}
\alias{get_slopes_and_mean}
\title{Get individual linear slope and mean values from longitudinal data}
\usage{
get_slopes_and_mean(
  data,
  id_col = "id",
  outcome_col = "outcome",
  time_col = "time"
)
}
\arguments{
\item{data}{a longitudinal data set}

\item{id_col}{the column name corresponding to the ID variable in data. Default = \code{"id"}}

\item{outcome_col}{the column name corresponding to the outcome variable in data. Default = \code{"outcome"}}

\item{time_col}{the column name corresponding to the time variable in data. Default = \code{"time"}}
}
\value{
a \code{tibble} with \code{id}, \code{performance_slope}, and \code{performance_mean} columns. This is what is used as inputs for subsequent \code{test_polynomial()} and \code{implement_polynomial()} function calls (and can therefore be directly piped in without storing intermediate output)
}
\description{
Get individual linear slope and mean values from longitudinal data
}
\examples{
# Create dummy data
n_participants = 5
n_timepoints = 3

set.seed(2222)

dummy <- data.frame(
  id = rep(1:n_participants, each = n_timepoints),
  timepoint = rep(1:n_timepoints, times = n_participants),
  performance = rnorm(n_participants * n_timepoints, mean = 0, sd = 1)
)

# Extract slopes and mean from dummy data
get_slopes_and_mean(
data = dummy,
id_col = "id",
time_col = "timepoint",
outcome_col = "performance"
)
}
