% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_polynomial.R
\name{test_polynomial}
\alias{test_polynomial}
\title{Test models with polynomials of differing orders to describe the relationship between linear rate of change and mean score in longitudinal data}
\usage{
test_polynomial(data, test_proportion = 1/3, max_order = 6)
}
\arguments{
\item{data}{output from \code{longpoly::get_slopes_and_mean()} (or any tibble with columns "id", "performance_slope", & "performance_mean")}

\item{test_proportion}{the proportion of participants to allocate to the test set. default = \code{1/3}}

\item{max_order}{the maximum order for polynomial models to be tested. Polynomial models for 1:\code{max_order} will be tested. default = \code{6}}
}
\value{
a list containing:
\enumerate{
\item \code{polynomial_results} — a tibble with columns recording the order of each polynomial tested with the corresponding proportion of variance explained (PVE) in test and train data, as well as the additional PVE in the test data for the increase in order for each
\item \code{train_ids}  — a character vector of ids allocated to the train dataset in model development
\item \code{test_ids}  — a character vector of ids allocated to the test dataset in model development
\item \code{scree_plot}  — a visualisation of the additional PVE for each polynomial. It is recommended that the order of the model be selected as that to the left of the ‘elbow’ (where the improvement plateaus).
}
}
\description{
Test models with polynomials of differing orders to describe the relationship between linear rate of change and mean score in longitudinal data
}
\examples{

# test_polynomial(slope_mean_tibble,
#                 test_proportion = 0.75,
#                 max_order = 5)
#
# # alternatively, the output from longpoly::get_slopes_and_mean() may be directly piped in, e.g.:
#
# get_slopes_and_mean(example_data,
#                     id_col = "id",
#                     outcome_col = "memory_test",
#                     time_col = "time") \%>\%
#   test_polynomial(.,
#                 test_proportion = 0.75,
#                 max_order = 5)

}
